requirements.txt
asgiref==3.7.2
certifi==2023.11.17
chardet==3.0.4
Django==4.2.2
django-filter==23.5
djangorestframework==3.14.0
docopt==0.6.2
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Markdown==3.5.1
num2words==0.5.13
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2023.3.post1
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
tzdata==2023.3
xlwt==1.3.0
